working

+++ classes.cpp
@@ -101,7 +101,7 @@ class Student{
 
 class Box{
     private:
-    int l, b, h;
+    long l, b, h;
 
     public:
     Box(){
